+++ CesiumVectorData/include/CesiumVectorData/GltfConverter.h
@@ -1,9 +1,13 @@
 #pragma once
 
+#include <Cesium3DTilesReader/ExtensionSchemaMaxarContentGeoJsonReader.h>
 #include <CesiumGeospatial/Ellipsoid.h>
 #include <CesiumGltf/Model.h>
+#include <CesiumGltf/Schema.h>
 #include <CesiumUtility/Result.h>
 #include <CesiumVectorData/GeoJsonDocument.h>
+
+#include <rapidjson/document.h>
 
 namespace CesiumVectorData {
 
@@ -11,6 +15,12 @@ namespace CesiumVectorData {
  * @brief The result of converting a GeoJSON document to glTF.
  */
 using ConverterResult = CesiumUtility::Result<CesiumGltf::Model>;
+
+/**
+ * @brief The result of converting a MAXAR_content_geojson schema to a glTF
+ * schema.
+ */
+using ConvertSchemaResult = CesiumUtility::Result<CesiumGltf::Schema>;
 
 /**
  * @brief Convert GeoJSON documents to glTF.
@@ -28,5 +38,13 @@ public:
   static ConverterResult convert(
       const GeoJsonDocument& geoJson,
       const CesiumGeospatial::Ellipsoid& ellipsoid);
+  /**
+   * @brief Convert MAXAR_content_geojson schema into a glTF schema.
+   * @param maxarSchema The schema extension object supplied with a GeoJson
+   * tileset
+   * @returns A result object that includes the converted schema and any errors.
+   */
+  static ConvertSchemaResult convertSchema(
+      const Cesium3DTiles::ExtensionSchemaMaxarContentGeoJson& maxarSchema);
 };
 } // namespace CesiumVectorData